Move Source defaults
