+++ obs_data/move_transition_left_corner.json
@@ -1,0 +1,35 @@
+{
+    "bounds": {
+        "x": 251.0,
+        "x_sign": " ",
+        "y": 234.0,
+        "y_sign": " "
+    },
+    "crop": {
+        "bottom": 0.0,
+        "bottom_sign": " ",
+        "left": 0.0,
+        "left_sign": " ",
+        "right": 0.0,
+        "right_sign": " ",
+        "top": 0.0,
+        "top_sign": " "
+    },
+    "duration": 300,
+    "pos": {
+        "x": -23.0,
+        "x_sign": " ",
+        "y": 345.0,
+        "y_sign": " "
+    },
+    "rot": 0.0,
+    "rot_sign": " ",
+    "scale": {
+        "x": 1.0,
+        "x_sign": " ",
+        "y": 1.0,
+        "y_sign": " "
+    },
+    "source": "Birb",
+    "transform_text": "pos: x -23.0 y 345.0 rot: 0.0 scale: x 1.000 y 1.000 crop: l 0 t 0 r 0 b 0"
+}
