+++ src/obs.rs
@@ -774,76 +774,6 @@ pub async fn follow(
 // Bootstrap / Create Things //
 // ========================= //
 
-pub async fn create_move_text_value_filter(
-    source: &str,
-    scene_item: &str,
-    filter_name: &str,
-    obs_client: &OBSClient,
-) -> Result<()> {
-    let base_settings =
-        move_transition::create_move_source_filter_settings(scene_item);
-    let new_settings =
-        move_transition::custom_filter_settings(base_settings, 1662.0, 13.0);
-
-    // "id": "move_value_filter",
-    // "mixers": 0,
-    // "monitoring_type": 0,
-    // "muted": false,
-    // "name": "OBS_Text",
-    // "prev_ver": 469827586,
-    // "private_settings": {},
-    // "push-to-mute": false,
-    // "push-to-mute-delay": 0,
-    // "push-to-talk": false,
-    // "push-to-talk-delay": 0,
-    // "settings": {
-    //     "custom_duration": true,
-    //     "duration": 300,
-    //     "filter": "",
-    //     "move_value_type": 4,
-    //     "setting_decimals": 1,
-    //     "setting_name": "text",
-    //     "setting_text": "we are working on getting\nfunctionality up",
-    //     "value_type": 4
-    // },
-    // "sync": 0,
-    let new_filter = obws::requests::filters::Create {
-        source,
-        filter: filter_name,
-        kind: "move_source_filter",
-        settings: Some(new_settings),
-    };
-    if let Err(err) = obs_client.filters().create(new_filter).await {
-        println!("Error Creating Filter: {filter_name} | {:?}", err);
-    };
-
-    Ok(())
-}
-
-pub async fn create_move_source_filters(
-    source: &str,
-    scene_item: &str,
-    filter_name: &str,
-    obs_client: &OBSClient,
-) -> Result<()> {
-    let base_settings =
-        move_transition::create_move_source_filter_settings(scene_item);
-    let new_settings =
-        move_transition::custom_filter_settings(base_settings, 1662.0, 13.0);
-
-    let new_filter = obws::requests::filters::Create {
-        source,
-        filter: filter_name,
-        kind: "move_source_filter",
-        settings: Some(new_settings),
-    };
-    if let Err(err) = obs_client.filters().create(new_filter).await {
-        println!("Error Creating Filter: {filter_name} | {:?}", err);
-    };
-
-    Ok(())
-}
-
 pub async fn create_outline_filter(
     source: &str,
     obs_client: &OBSClient,
@@ -1081,7 +1011,7 @@ pub async fn create_filters_for_source(
     }
 
     let filter_name = format!("Move_Source_Home_{}", source);
-    create_move_source_filters(
+    move_transition::create_move_source_filters(
         DEFAULT_SCENE,
         &source,
         &filter_name,
@@ -1165,7 +1095,7 @@ pub async fn create_filters_for_source(
 
     let filter_name = format!("Move_Source_{}", source);
 
-    create_move_source_filters(
+    move_transition::create_move_source_filters(
         DEFAULT_SCENE,
         &source,
         &filter_name,
